Updating README, minor alloc.c warnings, and updating actuator/radio usage of SPI.

+++ test/test_spi.c
@@ -2,8 +2,8 @@
 #include <stdio.h>
 
 void test_spi() {
-	int inst      = 1;
-	int ss_pins[1] = {43};
+	uint8_t inst = 1;
+	uint8_t ss_pins[1] = {43};
 
 	spi_init(inst, ss_pins, 1);
